07_Computer_Vision/04_EndtoEndProjects/services/backend_service/test/test_main.py
import pytest
from fastapi.testclient import TestClient
from app.main import app
from httpx import AsyncClient, ASGITransport


client = TestClient(app)

def test_read_root():
    response = client.get("/")
    assert response.status_code == 200
    assert response.json() == {"message": "Hello FastAPI"}


@pytest.mark.asyncio
async def test_ping():
    transport = ASGITransport(app=app)
    async with AsyncClient(transport=transport, base_url="http://test") as ac:
        response = await ac.get("/ping")
    assert response.status_code == 200
    assert response.json() == {"ping": "pong"}

